Add domain objects
The reservation model was incomplete. I added more models to make it complete.Because in the domain a reservation has a user and a restaurant associated.

However, I decided it is better to write down the domain and map the db for the listing reservations use case. I created a list with a couple of tasks to do that.