Set up http router with chi
I decided to use chi as a router because it is lightweight, idiomatic and composable. It will help organize the API as the project grow.

I set up the router with some essential middlewares
 * Log requests
 * Recover from panic
 * Format URL

I defined some basic routes to test if the web is available and to test if it recovers from a panic.
 * "/ping"
 * "/panic"