Implement reservations list endpoint
I created the endpoint to return a list of reservations.
Since I want to do something complete I also implemented the reservations
repository to fetch the reservations from the database.

Now the applications is starting to get format and I already have an use
case end to end implemented. I have to set up the tests structure
with a mocks package containing the objects that are going to be used among
all tests.