Get should run db migration from command line
Whenever I was starting the app I was seeding the db with mock data. This was causing the db to be polluted by too much data. In order to prevent this I created a should run db migration flag. By doing that I can pass a flag during start up to run the db migration. The detault is set to false because I don't need to run this often.

I added a TODO comment because eventually I will refactor this and create a proper seed and migration files for the DB.