Set up mock objects with testify and add missing tests for app
Initially I thought about create mock objects manually, but after thinking about it I thought it would make more sense to use testify and use a configuration similar to what I have used in other languages. Plus, I like to let is up to the client in a nice interface the definition of what the mock object should return.

In order to try out the mock object for the reservations repo I created the missing tests for the application layer.