Add missing tests for the reservations endpoint
I've added the missing tests for the endpoint. In the beggining I was struggling a bit to get it working, but I figured out it was not working because I was not calling the API init method to initialize the router. Since I did not want to do it all the time I move the Init() call to the API constructor.